Converted list of MapItemData into Map

+++ src/game/maps/mod.rs
@@ -37,6 +37,6 @@ pub struct MapObject {
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Position {
-    pub x: i64,
-    pub y: i64,
+    pub x: usize,
+    pub y: usize,
 }
